% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df}
\alias{df}
\title{Example DAM data}
\format{
A \code{data.frame} containing 28234 rows and 42 variables:
\describe{
\item{V1:V10}{Several columns of information generated by DAMScan. Only two columns here are of interest to us. Column V2 which has date information and column V3 which has time of day information.}
\item{V11:V42}{Activity counts for each of the 32 flies recorded in a single activity monitor.}
}
}
\usage{
df
}
\description{
A sample \code{data.frame} which has beam crossing data was collected every minute to facilitate activity and sleep analysis. Data is collected for wildtype flies under LD 12:12 and DD at 25 degree Celsius using the Drosophila Activity Monitoring (DAM) system, Trikinetics.
}
\keyword{datasets}
