% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilesAct.R
\name{profilesAct}
\alias{profilesAct}
\title{Compute and plot activity profiles}
\usage{
profilesAct(
  data,
  bin = 30,
  t.cycle = 24,
  average.type = "Both",
  rm.channels = c()
)
}
\arguments{
\item{data}{Input data file. The input for this function must be the output of the function binData(). See ??binData().}

\item{bin}{Intervals in which data are saved (in minutes). This defaults to 30. This value must be the same as that for binData().}

\item{t.cycle}{Define the period of the environmental cycle or a single day in hours. This defaults to 24. This value must be the same as that for binData().}

\item{average.type}{Define how the averaging must be done for computing profiles. There are 4 choices; i) "Flies": This will average over all the flies for the number of days of data that there are, and produce an averaged time-series, ii) "Days": This will average activity over all the days but for each fly and provide 32 averaged plots, iii) "Both": This will average over both flies and days and provide one composite average profile, iv) "None": This will not average and produce any plots; the output will be the same as the input file. This defaults to "Both".}

\item{rm.channels}{All the channels that users want to remove from their averaging. This must be a vector, i.e., channels must be separated by commas. For instance, if users choose to remove channels 1 to 5, 25 and 32, then the input should be either c(1,2,3,4,5,25,32) or c(1:5,25,32). This defaults to an empty vector, meaning no individuals are removed from analysis.}
}
\value{
Except when average.type = "None", a \code{list} with two items. When average.type = "None", input file is returned.
\description{
If average.type = "Days":
\item {Profiles}{
\description{
\item {ZT}{Column with ZT values.}
\item {I1:I32}{Data averaged over days for each of 32 flies.}
\item {ZT}{Column with ZT values.}
\item {I1:I32}{SEM (across days) for each of 32 flies.}
}
}
\item {Plot}{A \code{plotly} \code{htmlwidget} with the activity profiles in a 4-by-8 array.}
If average.type = "Flies":
\item {Profiles}{
\description{
\item {ZT}{Column with ZT values.}
\item {Mean}{Data averaged over all 32 flies for the entire duration of chosen days.}
\item {SEM}{SEM (across flies).}
}
}
\item {Plot}{A \code{plotly} \code{htmlwidget} with the activity time-series.}
If average.type = "Both":
\item {Profiles}{
\description{
\item {ZT}{Column with ZT values.}
\item {Mean}{Data averaged over all days and all 32 flies.}
\item {SEM}{SEM (across flies).}
}
}
\item {Plot}{A \code{plotly} \code{htmlwidget} with the activity profile.}
}

If average.type = "Days", "Flies" or "Both", this function returns a list with two elements. The first is a data frame called "Profiles", and the second is a plotly object called "Plot". If the average.type = "None", this function returns the input data set as is. When averaged over "Days", "Profiles" will be a data frame with 66 columns. Column 1 is ZT followed by 32 columns of activity data (averaged over days) for 32 flies. The 34th column is again ZT and the next 32 columns have the SEM values across cycles for each fly at that time point. The number of rows depends on the output bin size used in the binData function. When averaged over "Flies", "Profiles" will be a data frame with 3 columns. Column 1 is ZT, Column 2 is the averaged activity data over all flies, and Column 3 is the SEM across flies for that time point. The number of rows depends on the number of days of data and the output bin size used in the binData function. When averaged over "Both", "Profiles" will be a data frame with 3 columns. Column 1 is ZT, Column 2 is the averaged activity data over all flies and all days, and Column 3 is the SEM across flies for that time point. The number of rows depends on the output bin size used in the binData function.
}
\description{
Users can compute average profiles and visualise the same. Averages can be performed either over Flies, Days, Both or None. Except in the case of "None" the output of this function will be a list with two elements. One is the generated plot and the other is a table with the activity values and corresponding standard errors of the mean.
}
\examples{
td <- trimData(data = df, start.date = "19 Dec 20", start.time = "21:00",
n.days = 10, bin = 1, t.cycle = 24)
bd <- binData(td)
pro <- profilesAct(data = bd)
pro <- profilesAct(data = bd, bin = 60, t.cycle = 32,
average.type = "Days", rm.channels = c(1:5,25,32))
}
